Remap channels using an external mapping
This commit introduces --remap_file which can specify a channel
remapping. The format for the json is:

{
    "channel.1571434910068": {
      "original_call_sign": "KTVK",
      "remap_call_sign": "AZ Family",
      "original_channel": "3.1",
      "remap_channel": "3",
      "active": true
    },
    ...
}

original_call_sign and original_channel are unused, but serve as
documetation for the user and are required.

Note that --remap and --remap_file are mutually exclusive, although no
error is thrown currently when both options are specified. However,
--remap takes precedent. We might want to clean this up before this is
merged to `main`.

This commit also includes a sample remap.json, which should be cleanded
up and removed before this is merged into main.

Furthermore, this commit needs more documentation.

This is the first step to resolving issue #36